At runnable method (run), added all readed content lines at the common list of SensorsDataAggregator + Changed the list of SensorsDataAggregator to fix the race problem concurrent (read/write list) + Made some print method in order to see the results once all the threads are stopped
For avoid some concurrent problems, I changed LinkedList to LinkedBlockingQueue which is an optionally-bounded blocking queue based on linked nodes. This queue orders elements FIFO (first-in-first-out).

I was hesitating between the LinkedbBockingQueue and the ConcurrentLinkedQueue, but finally  I decided to use LinkedBlockingQueue because the LinkedBlockingQueue class implements the BlockingQueue interface, which provides the blocking nature to it.