Added a Semaphore object in order to make sure that when we write a new resultsFile, there are any thread adding new content to the common object where the data is stored
Added also a timer with a timertask which call the method of print a new results output file. This timertask is executed every 10 seconds.
Put a mutex (semaphore) in acquire when is trying to add new data to the common store object and when is trying to get all the data of it)