Added a new variable at SensorsDataAggregator which indicates if new information has been added to the queue (common object which stores all the read sensor data by the thread)
Before the program creates a new csv file every ten seconds independently of new data was added or not.
Now with this boolean variable, allows to the program that create a new csv only if a new sensor data is read.
Added some basic exception and moved the printwriter logic into a new class that is dedicated exclusively to it